requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
boto3==1.34.77
botocore==1.34.77
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.4
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.7.1
Django==4.2.11
django-celery-beat==2.6.0
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-filter==24.2
django-hosts==6.0
django-redis==5.4.0
django-rest-swagger==2.2.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
execnet==2.1.1
Faker==24.4.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.6
MarkupSafe==2.1.5
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.22
pytest==8.1.1
pytest-django==4.8.0
pytest-xdist==3.5.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
s3transfer==0.10.1
sentry-sdk==1.44.1
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==6.2
